refactor: autorefresh as a mixin
