+++ vue/mixins/index.js
@@ -5,6 +5,7 @@ export * from "./download";
 export * from "./file";
 export * from "./oauth";
 export * from "./part";
+export * from "./refresh";
 export * from "./ripe";
 export * from "./scroll";
 export * from "./utils";
